requirements.txt
cassandra-driver
Flask==1.1.2
